[Gold V] Title: 가로등, Time: 1164 ms, Memory: 239100 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/32069. 가로등/가로등.kt
@@ -1,21 +1,21 @@
 fun main() {
     val reader = System.`in`.bufferedReader()
-    var (numOfRoads, numOfLights, numOfAnswer) = reader.readLine().split(" ").map { it.toInt() }
+    var (numOfRoads, numOfLights, numOfAnswer) = reader.readLine().split(" ").map { it.toLong() }
     val answerNum = numOfAnswer
-    val roads = reader.readLine().split(" ").map { it.toInt() }
-    val setOfCountedRoad = mutableSetOf<Int>()
+    val roads = reader.readLine().split(" ").map { it.toLong() }
+    val setOfCountedRoad = mutableSetOf<Long>()
     val answer = mutableListOf<Int>()
-    var inc = 0
+    var inc = 0L
     while(numOfAnswer > 0){
 
         for(road in roads){
             if(!setOfCountedRoad.contains(road + inc) && road + inc <= numOfRoads){
-                answer.add(inc)
+                answer.add(inc.toInt())
                 setOfCountedRoad.add(road + inc)
                 numOfAnswer--
             }
             if(!setOfCountedRoad.contains(road - inc) && road - inc >= 0){
-                answer.add(inc)
+                answer.add(inc.toInt())
                 setOfCountedRoad.add(road - inc)
                 numOfAnswer--
             }
@@ -23,5 +23,5 @@ fun main() {
 
         inc++
     }
-    println(answer.take(answerNum).joinToString("\n"))
+    println(answer.take(answerNum.toInt()).joinToString("\n"))
 }
